Add readycheck.go operator util file for verifying deployments and daemonsets
