+++ pkg/virt-operator/util/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "client.go",
         "config.go",
+        "readycheck.go",
         "types.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-operator/util",
@@ -13,8 +14,10 @@ go_library(
         "//pkg/api/v1:go_default_library",
         "//pkg/controller:go_default_library",
         "//pkg/kubecli:go_default_library",
+        "//pkg/log:go_default_library",
         "//pkg/version:go_default_library",
         "//vendor/github.com/openshift/api/security/v1:go_default_library",
+        "//vendor/k8s.io/api/apps/v1:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
         "//vendor/k8s.io/client-go/tools/cache:go_default_library",
